requirements.txt
beautifulsoup4==4.12.3
readability-lxml==0.8.1
requests==2.31.0
elasticsearch==8.13.0
python-dotenv==1.0.1
openai==1.14.3
